LayoutTests/inspector/injected-script/observable-expected.txt

CONSOLE MESSAGE: line 25: [object Object]
Test that certain InjectedScriptSource operations are not observable by user code.


== Running test suite: InjectedScript.Observable
-- Running test case: InjectedScript.Observable:Array.prototype[Symbol.iterator]
PASS: Paused.
PASS: Resumed.
PASS: Array.prototype[Symbol.iterator] call count should be 1.
